+++ raycast-env.d.ts
@@ -13,31 +13,15 @@ type ExtensionPreferences = {}
 declare type Preferences = ExtensionPreferences
 
 declare namespace Preferences {
-  /** Preferences accessible in the `export-context` command */
-  export type ExportContext = ExtensionPreferences & {}
   /** Preferences accessible in the `capture-tana` command */
   export type CaptureTana = ExtensionPreferences & {}
-  /** Preferences accessible in the `ask-kai` command */
-  export type AskKai = ExtensionPreferences & {}
-  /** Preferences accessible in the `briefing` command */
-  export type Briefing = ExtensionPreferences & {}
-  /** Preferences accessible in the `commands` command */
-  export type Commands = ExtensionPreferences & {}
   /** Preferences accessible in the `create-tana-node` command */
   export type CreateTanaNode = ExtensionPreferences & {}
 }
 
 declare namespace Arguments {
-  /** Arguments passed to the `export-context` command */
-  export type ExportContext = {}
   /** Arguments passed to the `capture-tana` command */
   export type CaptureTana = {}
-  /** Arguments passed to the `ask-kai` command */
-  export type AskKai = {}
-  /** Arguments passed to the `briefing` command */
-  export type Briefing = {}
-  /** Arguments passed to the `commands` command */
-  export type Commands = {}
   /** Arguments passed to the `create-tana-node` command */
   export type CreateTanaNode = {}
 }